Refactoring: created Compass class
